fix(conf): make ES user/password inclusion conditional
This commit makes the elasticsearch user & password configuration lines conditional. I discovered a bug using AWS ElasticSearch service. They do not include a username/password. Omitting the FLUENT_ELASTICSEARCH_{USER,PASSWORD} environment variables on the DaemonSet generates blank lines in the generated files which end up as empty values. The result is HTTP authentication is used even when environment variables are not set. This is a workaround for possible issue in the elasticsearch output plugin.