v1.0: Initial migration to Fluentd 1.1.x
The following patch migrates the Fluentd version to v1.1.0 and perform
further adjustments in the configuration.

This patch removes supports for Alpine Linux: Fluentd in Kubernetes
may face a high memory usage due to memory fragmentation by glibc. In
order to avoid this situation we only keep a minimal-stable debian
image that's compatible with Jemalloc so we can get rid of the problem.

Fluentd must not be used in Kubernetes without Jemalloc.