Addressing some issues with request retries
The default request timeout is 5s, which can be too low for some systems such as Amazon ElastichSearch when sending across regions.  This can result in unnecessary retries, which then can lead to duplicate log entries.  To address this, the change adds support for setting a different request_timeout through environment variables.  It also uses elasticsearch_genid to generate document ids for ElasticSearch logs locally, which appears to be a best practice.

Signed-off-by: Joshua Garnett <josh.garnett@gmail.com>