created module for all AWS services
Signed-off-by: spensireli <spencerconklin92@gmail.com>

This creates a new version labeled "debian-aws".

It is pretty simple, it includes S3, Kinesis, and CloudWatch. This is not as light weight as choosing one of the versions for the specific service, however this would allow usage for multiple aws plugins.

This is important if you want to use FluentD to send to both CloudWatch & S3 or CloudWatch & Kinesis etc.

Often cost implications of CloudWatch make it beneficial to only send some data to CloudWatch and a large portion to S3.