Add parameter target_index_key in fluent.conf
The parameter target_index_key tell the plugin to find the index name to write to in the record under this key.

If the key is present in the record the value will be used as the index name. If it is not found then it will use index_name setting as configured. If the value of target_index_key if blank, then the plugin will not find the field in the record and use the index_name setting.


Suppose you have the following settings

target_index_key @target_index
index_name fallback