Add rexml
Because Ruby 3.x drops rexml from default gems, it's not installed by
default. aws-sdk-core requires one of XML parser gems but it's up to the
user to decide which one to actually use.
rexml is still "bundled gem" of Ruby 3.x so we choose to install it by
default.

Fix #1374

Signed-off-by: Takuro Ashie <ashie@clear-code.com>