Set default flush_interval to 5 for forward and syslog images
The default is 60, which results in a serious delay. Most of the other
backends set a lower interval. Since the work of forward and syslog is
minimal, flushing every 5 seconds should be safe.

The intervals can be set by env variables FLUENT_SYSLOG_FLUSH_INTERVAL
and FLUENT_FORWARD_FLUSH_INTERVAL respectively.

Signed-off-by: Orgad Shaneh <orgad.shaneh@audiocodes.com>