Change jemalloc opt-in for now
There is a known bug that combination with jemalloc and
fluent-plugin-systemd causes free(): invalid crash for
a long time. The problematic code is identified but the
root cause is not fixed yet.

There is a workaround for this - disable jemalloc explicitly.
LD_PRELOAD= stop to use jemalloc.

If you want to use jemalloc, set it via env like this:

 containers:
   - name: fluentd
     image: fluent/fluentd-kubernetes-daemonset:v1-debian-forward
     env:
       - name: K8S_NODE_NAME
         valueFrom:
           fieldRef:
             fieldPath: spec.nodeName
       - name:  FLUENT_FORWARD_HOST
         value: "REMOTE_ENDPOINT"
       - name:  FLUENT_FORWARD_PORT
         value: "18080"
       - name:  LD_PRELOAD
         value: "/usr/lib/libjemalloc.so.2"

Related issues:

https://github.com/fluent/fluentd-docker-image/issues/378
https://github.com/fluent/fluent-package-builder/issues/369
https://github.com/fluent-plugins-nursery/fluent-plugin-systemd/issues/110
https://github.com/ledbettj/systemd-journal/issues/93
https://github.com/fluent/fluentd-kubernetes-daemonset/issues/1478

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>